[FLINK-17426][blink-planner] DynamicSource supports SupportsLimitPushDown.
