+++ flink-table/flink-table-planner-blink/src/test/scala/org/apache/flink/table/planner/runtime/batch/sql/LimitITCase.scala
@@ -18,106 +18,30 @@
 
 package org.apache.flink.table.planner.runtime.batch.sql
 
-import org.apache.flink.table.api.TableSchema
-import org.apache.flink.table.api.ValidationException
-import org.apache.flink.table.planner.runtime.utils.BatchTestBase
-import org.apache.flink.table.planner.runtime.utils.TestData._
-import org.apache.flink.table.planner.utils.TestLimitableTableSource
+import org.apache.flink.table.planner.factories.TestValuesTableFactory
+import org.apache.flink.table.planner.runtime.utils.{BatchTestBase, TestData}
+import org.apache.flink.table.planner.runtime.utils.TestData.{data3, nullablesOfData3, type3}
 
-import org.junit._
-
-class LimitITCase extends BatchTestBase {
-
-  @Before
+class LimitITCase extends LegacyLimitITCase {
   override def before(): Unit = {
-    super.before()
+    BatchTestBase.configForMiniCluster(conf)
     registerCollection("Table3", data3, type3, "a, b, c", nullablesOfData3)
 
-    TestLimitableTableSource.createTemporaryTable(
-      tEnv, data3, new TableSchema(Array("a", "b", "c"), type3.getFieldTypes), "LimitTable")
-  }
-
-  @Test
-  def testOffsetAndFetch(): Unit = {
-    checkSize(
-      "SELECT * FROM Table3 OFFSET 2 ROWS FETCH NEXT 5 ROWS ONLY",
-      5)
-  }
-
-  @Test
-  def testOffsetAndLimit(): Unit = {
-    checkSize(
-      "SELECT * FROM Table3 LIMIT 10 OFFSET 2",
-      10)
-  }
-
-  @Test
-  def testFetch(): Unit = {
-    checkSize(
-      "SELECT * FROM Table3 FETCH NEXT 10 ROWS ONLY",
-      10)
-  }
-
-  @Test
-  def testFetchWithLimitTable(): Unit = {
-    checkSize(
-      "SELECT * FROM LimitTable FETCH NEXT 10 ROWS ONLY",
-      10)
-  }
-
-  @Test
-  def testFetchFirst(): Unit = {
-    checkSize(
-      "SELECT * FROM Table3 FETCH FIRST 10 ROWS ONLY",
-      10)
-  }
-
-  @Test
-  def testFetchFirstWithLimitTable(): Unit = {
-    checkSize(
-      "SELECT * FROM LimitTable FETCH FIRST 10 ROWS ONLY",
-      10)
-  }
-
-  @Test
-  def testLimit(): Unit = {
-    checkSize(
-      "SELECT * FROM Table3 LIMIT 5",
-      5)
-  }
-
-  @Test
-  def testLimit0WithLimitTable(): Unit = {
-    checkSize(
-      "SELECT * FROM LimitTable LIMIT 0",
-      0)
-  }
-
-  @Test
-  def testLimitWithLimitTable(): Unit = {
-    checkSize(
-      "SELECT * FROM LimitTable LIMIT 5",
-      5)
-  }
-
-  @Test(expected = classOf[ValidationException])
-  def testTableLimitWithLimitTable(): Unit = {
-    Assert.assertEquals(
-      executeQuery(tEnv.from("LimitTable").fetch(5)).size,
-      5)
-  }
-
-  @Test
-  def testLessThanOffset(): Unit = {
-    checkSize(
-      "SELECT * FROM Table3 OFFSET 2 ROWS FETCH NEXT 50 ROWS ONLY",
-      19)
-  }
-
-  @Test
-  def testLessThanOffsetWithLimitSource(): Unit = {
-    checkSize(
-      "SELECT * FROM LimitTable OFFSET 2 ROWS FETCH NEXT 50 ROWS ONLY",
-      19)
+//    TestLimitableTableSource.createTemporaryTable(
+//      tEnv, data3, new TableSchema(Array("a", "b", "c"), type3.getFieldTypes), "LimitTable")
+    val myTableDataId = TestValuesTableFactory.registerData(TestData.data3)
+    val ddl =
+      s"""
+         |CREATE TABLE LimitTable (
+         |  a int,
+         |  b bigint,
+         |  c string
+         |) WITH (
+         |  'connector' = 'values',
+         |  'data-id' = '$myTableDataId',
+         |  'bounded' = 'true'
+         |)
+       """.stripMargin
+    tEnv.executeSql(ddl)
   }
 }
